devices/radio/RASPi/requirements.txt
# Raspberry Pi runtime dependencies for the Radio device

httpx>=0.27.2
openai>=1.65.0
pydantic>=2.0
PyYAML>=6.0.2
python-dotenv>=1.0.1
gpiozero>=2.0
